requirements.txt
arrow==0.12.1
backports.functools-lru-cache==1.5
binaryornot==0.4.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
Django==1.8
djangorestframework==3.2.0
future==0.16.0
idna==2.6
MarkupSafe==1.0
python-dateutil==2.7.0
requests==2.18.4
six==1.11.0
urllib3==1.22
whichcraft==0.4.1
